test(server): merge the active-run watchdog module files to cut per-file cost
Each test file pays a fixed transform-and-import cost before its first
case runs. Three module test files paid that cost three times; one
merged file pays it once. The merge also collapses four near-identical
RunSnapshot factories into one and three near-identical
WatchdogRunReader factories into one, with no case or assertion lost.

The domain, application, and adapters production files merge the same
way, one file per layer group, so a reader still sees the layer split
through the file's describe blocks and its own internal sections.
